requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.3
cmake==4.1.0
comm==0.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==2.15.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
easydict==1.13
emoji==2.14.1
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2023.10.0
h11==0.16.0
h5py==3.8.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.2.2
Levenshtein==0.21.0
lit==18.1.8
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.1.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.15
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
overrides==7.7.0
packaging==25.0
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.4
peft==0.7.1
pexpect==4.9.0
pillow==11.3.0
pip==25.1
platformdirs==4.3.7
prometheus_client==0.22.1
prompt-toolkit==3.0.43
propcache==0.3.2
protobuf==6.32.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-Levenshtein==0.21.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.14.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.12.2
Send2Trash==1.8.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.2.0
stanza==1.5.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.5.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.28.1
triton==2.0.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
